pyproject.toml
[tool.poetry]
name = "ncwrapper"
version = "0.1.0"
description = ""
authors = ["trojvn <mrx.trojan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
httpwrapper = {git = "https://github.com/trojvn/httpwrapper.git"}


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"
python-dotenv = "^1.1.0"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
